requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-multipart
pyjwt
passlib[bcrypt]
email-validator
celery
redis
cloudinary